requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
fuzzywuzzy==0.18.0
kombu==5.4.2
Markdown==3.7
npm==0.1.1
openpyxl==3.1.5
optional-django==0.1.0
phonenumbers==8.13.52
pillow==11.0.0
prompt_toolkit==3.0.50
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
